🎁 Convert more URIs to string values
This commit will make it so the subject URIs on collections are
converted to string values.  Also, contributor was not being turned into
a string, even though it doesn't show up in the collection's show page,
it does show up in the facets in a catalog search.  Same thing with the
language field on both works and collection.  Interestingly, currently
there doesn't seem to be a way to set the language field in the
Collection form.  This possibly is a bug but admins can still set it
through Bulkrax.

Ref:
- https://github.com/notch8/utk-hyku/issues/592