🎁 Show thumbnail if attachments are institution
This commit will add logic to only show the thumbnail and not the UV
when the work has attachments that are set to institution visibility to
public users.  Also added minor styling so the thumbnail is centered on
the show page and not aligned to the left.  This allows for the
thumbnail and metadata for the works in the Archivision to be publicly
visible.  The attachments and file sets would need to be set to
"Institution" (authenticated).

Ref:
- https://github.com/notch8/utk-hyku/issues/647