🐛 Adds "more" option to facets with 5+ items
This commit ensures all facet options have a "more" option when there
are over five items in the facet. This is to prevent the facet list from
becoming too long.

Blacklight will look to see if the facet has at least limit + 1 items
when determining if the "more" option should be displayed.

The Solr fields with the suffix `_sim` should be used in the
`config.add_facet_field` ad can be renamed using the label attribute.

Ref:
- https://github.com/notch8/utk-hyku/issues/656