🧹 Improve aggregating solr fields
In UTK's m3 profile (looking at the YAML), we have something that looks
like: `properties[property_name]['mappings']['blacklight']`.  The value
is used to indicate what properties should be aggregated into what solr
field.  For example, `properties.photographer.mappings.blacklight` is
`creator_sim`.  That means when we have the facet drop down in the
catalog page, the photographer name should be in the `Creator` facet.
Previously, we created a method called SolrDocument.creator_fields,
which returned all properties that has `creator_sim` in the
`...mappings.blacklight`.  This commit will create a way for these
fields (only dealing with sim) to be auto generated and indexed
accordingly.