🧹 Accounting for the `license` property
This commit will add in consideration for license, which is also a local
authority.  Logic has been added to check locally first before going out
to Creative Commons.  Also, fix a small bug where the Rights Statement
and License was not displaying on the work show page.  Lastly, both
Rights Statement and License are not faceted via the CatalogController.
We will need to remove the `facetable` from the properties to stop
Rights Statement and License rendering separately.

Ref:
- https://github.com/notch8/utk-hyku/issues/637