🧹 Change "Homepage" to "Link to item"
This commit wil remove the previous attempt at solving this go for
directly editing the compiled javascript files.  This is a quick
solution so I'm definitely opened to a better one.

There is now i18n support for CloverIIIF however I didin't see anything
that would affect this particular request for changing the "Homepage"
label.

Ref:
- https://github.com/notch8/utk-hyku/issues/659