🐛 Fix File Manager save button
The file manager save button was still disabled even after changes were
made to the files.  This was caused by a selector error because we are
running jQuery 3.  That change was made in Hyrax.  The other change was
to add a decorator for the `Hyrax::Actors::ApplyOrderActor` to compact
and nil objects in the ordered_member_ids.  Why and how it got in there
in the first place?  I haven't been able to solve that part.  But, after
saving it with this override, it seemed to fix the issue and I no longer
see the nil in ordered_member_ids.

Ref:
- https://github.com/notch8/utk-hyku/issues/580
- https://github.com/samvera/hyrax/commit/a4890955954a6d5d824a4c8786374f9ff54355e3