🎁 Add a proxy for VTT transcript files
Since we're in an S3 backed Fedora environment, CloverIIIF makes a call
to S3 to get the Video Text Tracks (VTT) content, however it gets
blocked because of CORS issues.  This commit gets around that by adding
a new route and indexing the WEBVTT content onto the file set which gets
sent back to CloverIIIF instead so we don't have to deal with S3 at all.

Ref:
- https://github.com/notch8/utk-hyku/issues/442