🧹 Filter out unwanted work types from OAI feed
This commit will add an override for the Blacklight OAI Provider gem to
add a fq for unwanted work types.  We're keying off of the
IiifPrint::SearchBuilder to keep things consistent.  At the moment, we
don't want Attachment work types to show int he Catalog searches nor in
the OAI feed.

Ref:
- https://github.com/notch8/utk-hyku/issues/680