🎁 Add a database model for caching uri's
The issue we were facing was that LoC and possibily other authorities
will rate limit the API calls.  This will result in a lot of URIs being
not resolved and fails to index correctly on the SolrDocument.  This
local cache will save URI's so we can look here first then look for it
remotely.