🎁 Add URI caching to URI lookup
This commit will make our URI lookup to check the UriCache first before
reaching out to the remote authority.  If a URI is already cached, then
it will return that value and move on.  If it is not already cached,
then it will create a database record of it for next time.