🐛 Account for concurrency
Because we have multiple workers running we may end up with a race
condition where worker_1 will check the cache and doesn't find the uri
so it will fetch from remote.  Then while that's happening worker_2 also
checks the cache for the same URI and also doesn't find it, so it it'll
fetch from remote as well.  Meanwhile worker_1 finishes and creates the
UriCache record.  Then worker_2 finishes and tries to create the same
UriCache record but end up with a `ActiveRecord::RecordNotUnique` error.
To account for this we add a rescue to look it up in the cache and move
on.