🎁 Adjust collection searching
UTK has an Attachment work type that serves as a proxy for the file set.
When Attachments are added to works they are not added to the collection
that the work is added to.  This causes an inconsistency between the
catalog search and the collection search.  Attachments are found in the
catalog search but return their parent works and in a collection search
Attachments are not found at all since they are not directly apart of
the the collection.

This commit will alter the collection search builder so that the
Attachment that is a member of a work (which is the member of the
collection will) be returned.  However, we don't want to just return the
Attachment object, we return it's parent as a result.  This is how the
catalog search behaves so we borrowed some of its query params.

Also, the `abstract` field is what we're actually using instead of
`description`, so we are adding `abstract_tesim` to do the default
solr params qf.

Lastly, `member_of_collection` was being added again to the
`default_processor_chain` which is not need since it's being added to
in Hyrax.  We remove the duplication here so it doesn't get called
twice.