🎁 Add more metadata and search results
Previously, we were only given the thumbnail and the title of the work
on the collection's show page.  Now we have the abstract and date
issued/date created.  In addition, if we are doing a search and get a
hit, there will be a search result indicating with child work (usually
the Attachment) had the search term.

Ref:
- https://github.com/notch8/utk-hyku/issues/750