🐛 Fix not being able to delete works with Bulkrax
This commit will add an override for the #find method in the
`Bulkrax::FactoryClassFinder`.  We we getting the name set to the
default type of `GenericWork` because it tries to look in the
`parsed_metadata` for the `model` but if we don't have any
`parsed_metadata` at that point then it does the fallback.

Ref:
- https://github.com/notch8/utk-hyku/issues/784