🎁 Attachment and FileSet objects get destroyed
This commit will ensure that when a work is deleted, its Attachment
objects are also deleted and will then delete all the Attachments'
FileSets.  Bulkrax was not set up to delete correctly because it wasn't
using the actor stack.  This commit will override Bulkrax to properly go
through the actor stack for delete actions.

Ref:
- https://github.com/notch8/utk-hyku/issues/784